+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 Unreleased
 
+* Fixed issued when setting session cookie (svenkubiak)
 * Version bumps (svenkubiak)
   * mongo-driver-sync 5.7.0 -&gt; 5.8.0